% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosqcontrol.R
\docType{package}
\name{mosqcontrol}
\alias{mosqcontrol}
\alias{mosqcontrol-package}
\title{mosqcontrol: Mosquito Control Resource Optimization}
\description{
This project aims to make an accessible model for mosquito control
resource optimization. The model uses data provided by users to estimate
the mosquito populations in the sampling area for the sampling time
period, and the optimal time to apply a treatment or multiple treatments.
}
\author{
\strong{Maintainer}: Travis Byrum \email{tbyrum@terpmail.umd.edu}

Authors:
\itemize{
  \item Jeff Demers \email{jdemeripi@gmail.com}
  \item Anshuman Swain \email{answain@terpmail.umd.edu}
  \item Sharon Bewick \email{sharon_bewick@hotmail.com}
  \item William Fagan \email{bfagan@umd.edu}
}

}
