% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product-values.R
\name{P_vals}
\alias{P_vals}
\title{P_vals}
\usage{
P_vals(z, rho, Npulse, tau, k)
}
\arguments{
\item{z}{Numeric}

\item{rho}{Numeric}

\item{Npulse}{Numeric}

\item{tau}{Numeric}

\item{k}{Numeric}
}
\description{
\code{P_vals} returns "P" product.
}
\details{
This function computes the "P" product and exponentials for a given k.
}
\keyword{internal}
