% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product-values.R
\name{product_values}
\alias{product_values}
\title{product_values}
\usage{
product_values(z, rho, n_pulse, tau, k)
}
\arguments{
\item{z}{Numeric}

\item{rho}{Numeric}

\item{n_pulse}{Numeric}

\item{tau}{Numeric}

\item{k}{Numeric}
}
\description{
\code{product_values} returns "P" product.
}
\details{
This function computes the "P" product and exponentials for a given k.
}
