% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-values.R
\name{Q_vals}
\alias{Q_vals}
\title{Q_vals}
\usage{
Q_vals(z, rho, Npulse, tau, j, jj)
}
\arguments{
\item{z}{Numeric}

\item{rho}{Numeric}

\item{Npulse}{Numeric}

\item{tau}{Numeric}

\item{j}{Matrix}

\item{jj}{Matrix}
}
\description{
\code{Q_vals} returns "Q" product.
}
\details{
This function computes the "Q" product and exponentials with the fourier
modes for a given k
}
\keyword{internal}
