% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-values.R
\name{q_values}
\alias{q_values}
\title{q_values}
\usage{
q_values(z, rho, n_pulse, tau, j, jj)
}
\arguments{
\item{z}{Numeric}

\item{rho}{Numeric}

\item{n_pulse}{Numeric}

\item{tau}{Numeric}

\item{j}{Matrix}

\item{jj}{Matrix}
}
\description{
\code{q_values} returns "Q" product.
}
\details{
This function computes the "Q" product and exponentials with the fourier
modes for a given k
}
