% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse.R
\name{sum_one_pulse}
\alias{sum_one_pulse}
\title{sum_one_pulse}
\usage{
sum_one_pulse(z, rho, mu, tau, kmax, modes)
}
\arguments{
\item{z}{Numeric}

\item{rho}{Numeric}

\item{mu}{Numeric}

\item{tau}{Numeric}

\item{kmax}{Numeric}

\item{modes}{Numeric}
}
\description{
\code{sum_one_pulse} returns output pulse.
}
\details{
This function computes the fourier sum for 1 pulse.
}
\keyword{internal}
