% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{uperm}
\alias{uperm}
\title{uperm}
\usage{
uperm(d)
}
\arguments{
\item{d}{Vector}
}
\description{
\code{uperm} returns permutation matrix.
}
\details{
For a given list of numbers, this function outputs a matrix, where each row
is a unique permutation of the list.
}
\examples{
uperm(c(1, 2))
}
